Changes to the MGRX tests: renderize paths by its own, using integer arithmetic do not calculate bezier curves if they are out of the clip area rsvg2 prints calculated image dimensions, so you can edit the svg file if the reported dimensions are bad values rsvg2 checks for overflow when zooming

+++ mgrx/rendmgrx.h
@@ -35,8 +35,8 @@ typedef struct {
     int mode;
     int adj;
     double zoom;
-    int xdespl;
-    int ydespl;
+    double xdespl;
+    double ydespl;
     GrColor bg;
 } GrSVGDrawMode;
